doesnt read app.properties
and city tests run in bunch fail
