+++ MishpahugHaifa/src/main/resources/application.properties
@@ -4,3 +4,6 @@ spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL57Dialect
 spring.datasource.url=jdbc:mysql://localhost:3306/mishpahughaifa?serverTimezone=Asia/Jerusalem&useSSL=false
 spring.datasource.username=root 
 spring.datasource.password=root
+
+spring.jpa.show-sql=false
+hibernate.show_sql=false
