+++ MishpahugHaifa/src/test/java/Application/entities/country/CountryTest.java
@@ -1,20 +1,22 @@
 package Application.entities.country;
 
-import Application.entities.CountryEntity;
-import Application.repo.CountryRepository;
+import java.io.BufferedReader;
+import java.io.FileReader;
+import java.io.IOException;
+import java.util.List;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.io.BufferedReader;
-import java.io.FileReader;
-import java.io.IOException;
-import java.util.List;
+import Application.entities.CountryEntity;
+import Application.repo.CountryRepository;
 
 @RunWith(SpringRunner.class)
 @DataJpaTest
