+++ MishpahugHaifa/src/test/resources/application.properties
@@ -1,0 +1,12 @@
+jdbc.driverClassName=org.h2.Driver
+jdbc.url=jdbc:h2:mem:nodb;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false;
+
+jdbc.user=sa
+jdbc.pass=
+ 
+hibernate.dialect=org.hibernate.dialect.H2Dialect
+hibernate.hbm2ddl.auto=create
+
+
+spring.jpa.show-sql=false
+hibernate.show_sql=false
