+++ MishpahugHaifa/src/test/resources/data.csv
@@ -1,0 +1,40 @@
+title_en
+Korazim
+Lehavim
+Liman
+Livnim
+Maon
+Ma'agalim
+Ma'ale Amos
+Ma'anit
+Masua
+Matan
+Menora
+Menuha
+Mitzpa
+Mitav
+Moreshet
+Naan
+Naham
+Neve Ilan
+Nir Oz
+Beit Nir
+Ben Ami
+Beniya
+Berosh
+Biriya
+Buq'ata
+Ein Gev
+Ein HaEmek
+Eli-Ad
+Ezer
+Gan Haim
+Ganei Tal
+Gan Or
+Gezer
+Geva
+Gevat
+Gia
+Gonen
+Hamra
+Hanita
