+++ MishpahugHaifa/src/test/resources/data_country.csv
@@ -1,0 +1,234 @@
+Iraq
+Georgia
+Mexico
+Belize
+Aruba
+Bhutan
+Denmark
+Mongolia
+Monaco
+Haiti
+Ireland
+Mali
+Congo
+Guam
+Angola
+Cuba
+Laos
+Libya
+Ghana
+Gambia
+Belgium
+Benin
+Djibouti
+Cyprus
+Moldova
+Latvia
+Italy
+China
+Zambia
+Guinea
+Macau
+India
+Kenya
+Malawi
+Guyana
+Yemen
+Qatar
+Malaysia
+Comoros
+Mauritius
+Myanmar
+Germany
+Nauru
+Nepal
+Vatican
+Uganda
+Estonia
+Singapore
+Jamaica
+Niger
+Russia
+Nicaragua
+Ethiopia
+Serbia
+Gabon
+Turkey
+Sudan
+Bahamas
+Niue
+Togo
+Croatia
+Bahrain
+Thailand
+Somalia
+Palau
+Tokelau
+France
+Finland
+Israel
+Sri Lanka
+Samoa
+Senegal
+Swaziland
+Bermuda
+Slovenia
+Taiwan
+Chile
+USA
+Fiji
+Tuvalu
+Norway
+Rwanda
+Chad
+Isle of Man
+Oman
+Panama
+Peru
+Réunion
+Ecuador
+Kyrgyzstan
+Iran
+Seychelles
+Morocco
+Cape Verde
+Dominica
+Cayman Islands
+Zimbabwe
+Saudi Arabia
+Grenada
+Montserrat
+Malta
+Tajikistan
+Uruguay
+Kazakhstan
+Tunisia
+Canada
+East Timor
+Anguilla
+Tonga
+Indonesia
+Sint Maarten
+Mozambique
+Bulgaria
+Jordan
+Greece
+Liberia
+Slovakia
+Costa Rica
+Romania
+Namibia
+Lesotho
+Macedonia
+Hong Kong
+Lebanon
+Nigeria
+Tanzania
+Trinidad and Tobago
+Guatemala
+Pakistan
+Côte d`Ivoire
+Japan
+Azerbaijan
+Gibraltar
+Western Sahara
+Martinique
+Uzbekistan
+Hungary
+Luxembourg
+Philippines
+Brazil
+Spain
+Wallis and Futuna
+Austria
+Mauritania
+Afghanistan
+Lithuania
+Belarus
+Switzerland
+New Caledonia
+Curaçao
+Faroe Islands
+Turkmenistan
+Bolivia
+Papua New Guinea
+Solomon Islands
+United Kingdom
+Madagascar
+Equatorial Guinea
+Cameroon
+Brunei Darussalam
+Ukraine
+Montenegro
+Sierra Leone
+El Salvador
+New Zealand
+Syria
+French Guiana
+Australia
+Andorra
+Venezuela
+United Arab Emirates
+San Marino
+Guinea-Bissau
+Bangladesh
+Botswana
+Albania
+Greenland
+South Sudan
+North Korea
+Norfolk Island
+US Virgin Islands
+Honduras
+Dominican Republic
+Guadeloupe
+Kuwait
+Central African Republic
+Northern Mariana Islands
+Czech Republic
+Netherlands
+British Virgin Islands
+São Tomé and Príncipe
+Burundi
+Suriname
+Cook Islands
+Saint Helena
+Saint Kitts and Nevis
+Falkland Islands
+Vanuatu
+South Korea
+South Africa
+Vietnam
+Congo, Democratic Republic
+Egypt
+Kiribati
+Puerto Rico
+Eritrea
+French Polynesia
+Turks and Caicos Islands
+Paraguay
+Poland
+Saint Vincent and the Grenadines
+Palestine
+Svalbard and Jan Mayen
+Algeria
+Barbados
+Burkina Faso
+Colombia
+Iceland
+Maldives
+Marshall Islands
+Micronesia
+Pitcairn Islands
+Saint Pierre and Miquelon
+Sweden
+American Samoa
+Antigua and Barbuda
+Argentina
+Armenia
+Cambodia
+Bonaire, Sint Eustatius and Saba
+Portugal
+Saint Lucia
+Bosnia and Herzegovina
+Liechtenstein
